refactor: move household survey endpoints in endpoints submodule
