+++ tests/integration/test_api_client.py
@@ -1,7 +1,6 @@
 import pandas as pd
 import pytest
 
-
 pytestmark = pytest.mark.integration
